moving new tests to unit tests
